New test that should work with if statement and remove old one test.
