--- sycl/test/feature-tests/inline-asm/asm_temp2.cpp
@@ -1,64 +1,0 @@
-// UNSUPPORTED: cuda
-// REQUIRES: gpu,linux
-// RUN: %clangxx -fsycl %s -DINLINE_ASM -o %t.out
-// RUN: %t.out
-// RUN: %clangxx -fsycl %s -o %t.ref.out
-// RUN: %t.ref.out
-
-#include "include/asmhelper.h"
-#include <CL/sycl.hpp>
-#include <stdlib.h> 
-#include <time.h>
-
-using dataType = cl::sycl::cl_int;
-
-template <typename T = dataType>
-struct KernelFunctor : WithOutputBuffer<T> {
-  bool swtch_feild;
-  KernelFunctor(size_t problem_size, bool swtch) : WithOutputBuffer<T>(problem_size), swtch_feild(swtch) {}
-  
-  void operator()(cl::sycl::handler &cgh) {
-    auto C = this->getOutputBuffer().template get_access<cl::sycl::access::mode::write>(cgh);
-    bool sw = 0;
-    cgh.parallel_for<KernelFunctor<T>>(
-        cl::sycl::range<1>{this->getOutputBufferSize()}, [=](cl::sycl::id<1> wiID) [[cl::intel_reqd_sub_group_size(16)]] {
-          volatile int output = 0;
-#if defined(INLINE_ASM) && defined(__SYCL_DEVICE_ONLY__)
-          asm volatile("cmp (M1,16) %1(0, 0)<1> 0x0:d\n"
-                       "je l1%=\n"
-                       "mov (M1,16) %0(0,0)<1> 0x7:d\n"
-                       "jmp l2%=\n"
-                       "l1%=:\n"
-                       "mov (M1,16) %0(0,0)<1> 0x8:d\n"
-                       "l2%=:"
-                              : "=rw"(output)
-                              : "rw"(sw));
-#else
-          if (sw)
-          {
-            output = 7;
-          }
-          else
-          {
-            output = 8;
-          }
-#endif
-          C[wiID] = output;
-        });
-  }
-};
-
-int main() {
-  srand(time(NULL));
-  bool swtch = bool(rand() % 2);
-  KernelFunctor<> f(DEFAULT_PROBLEM_SIZE, swtch);
-  
-  if (!launchInlineASMTest(f))
-    return 0;
-  /*else if (swtch && verify_all_the_same(f.getOutputBufferData(), 7))
-    return 0;*/
-  else if (/*!swtch && */verify_all_the_same(f.getOutputBufferData(), 8))
-    return 0;
-
-  return 1;
-}
